rsl: Re-introduce the SACCH deactivation marker
GSM 08.58 is not clear if the SACCH Deactivate is an idempotent operation,
it has exposed a bug in my osmo-bts channel release re-work with the queue
and I assume the nanoBTS will not deal with it any better.

In my tests it was possible to get a N200+1 times error during the channel
release procedure which lead to releasing the SACCH again. When this timeout
occurs it is very difficult to find out if the SACCH was already deactivated.

Re-introduce the usage of the sacch_deact flag in the lchan. When the release
is triggered by higher levels the SACCH should be active and we check for this
with a log message, in case of the error handling we check if the SACCH was
de-activated before de-activating it even if we are requested to do it. This
way we know the SACCH will be de-activated at most once.